car_service/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
